requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
docutils==0.22.3
filetype==1.2.0
glcontext==3.0.0
idna==3.11
Kivy==2.3.1
Kivy-Garden==0.1.5
llvmlite==0.46.0
moderngl==5.12.0
multipledispatch==1.0.0
numba==0.63.1
numpy==2.3.5
Pygments==2.19.2
pyrr==0.10.3
Pillow>=10.0
PyYAML==6.0.3
requests==2.32.5
sgp4>=2.21
skyfield>=1.46
urllib3==2.6.2
